Lecture Complete

+++ app/Http/Controllers/Api/TweetController.php
@@ -3,10 +3,12 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
+use App\Http\Requests\StoreTweetRequest;
+use App\Http\Requests\UpdateTweetRequest;
 use App\Models\Tweet;
-
 use App\Services\TweetService;
+
+use Illuminate\Support\Facades\Gate;
 
 class TweetController extends Controller
 {
@@ -21,6 +23,9 @@ class TweetController extends Controller
      */
     public function index()
     {
+
+        Gate::authorize('viewAny', Tweet::class);
+
         $tweets = $this->tweetService->allTweets();
         return response()->json($tweets);
     }
@@ -28,11 +33,10 @@ class TweetController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(StoreTweetRequest $request)
     {
-        $request->validate([
-        'tweet' => 'required|max:255',
-        ]);
+        Gate::authorize('create', Tweet::class);
+
         $tweet = $this->tweetService->createTweet($request);
         return response()->json($tweet, 201);
     }
@@ -42,17 +46,18 @@ class TweetController extends Controller
      */
     public function show(Tweet $tweet)
     {
+        Gate::authorize('view', $tweet);
+
         return response()->json($tweet);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Tweet $tweet)
+    public function update(UpdateTweetRequest $request, Tweet $tweet)
     {
-        $request->validate([
-        'tweet' => 'required|string|max:255',
-        ]);
+        Gate::authorize('update', $tweet);
+
         $updatedTweet = $this->tweetService->updateTweet($request, $tweet);
     
         return response()->json($updatedTweet);
@@ -63,6 +68,8 @@ class TweetController extends Controller
      */
     public function destroy(Tweet $tweet)
     {
+        Gate::authorize('delete', $tweet);
+
         $this->tweetService->deleteTweet($tweet);
         return response()->json(['message' => 'Tweet deleted successfully']);
     }
